// src/lib/api/generated/model/selfProfileSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export interface SelfProfileSchema {
  campus?: string | null;
  college?: string | null;
  department?: string | null;
  designation?: string | null;
  display_name: string;
  is_head_guidance?: boolean | null;
  lifecycle_status?: string | null;
  profile_type?: string | null;
  program?: string | null;
  role: string;
  student_number?: string | null;
  user_id: number;
  year_level?: string | null;
}